+++ src/earley/traversal.rs
@@ -1,251 +1,287 @@
-use super::grammar::{
-    GrammarMatching, Production, ProductionId, ProductionMatch, ProductionMatching, TermMatch,
-};
-use super::input_range::{InputRange, InputRangeOffset};
+use super::{grammar::ProductionId, InputRange, Production};
 use crate::{
     append_vec::{append_only_vec_id, AppendOnlyVec},
-    tracing, Term,
+    Term,
 };
-use std::collections::{HashMap, HashSet, VecDeque};
-use std::rc::Rc;
+use std::collections::HashMap;
 
-/// The three main steps of the "Earley" parsing algorithm
-#[derive(Debug)]
-pub(crate) enum EarleyStep<'gram> {
-    /// If the next [`crate::Term`] is [`crate::Term::Nonterminal`] then "predict" more [`Traversal`]s
-    Predict(&'gram Term),
-    /// If the next [`crate::Term`] is [`crate::Term::Terminal`] then "scan" input text
-    Scan(&'gram String),
-    /// If the [`ProductionMatching`] has no unmatched [`crate::Term`]s then "complete" pending [`Traversal`]s
-    Complete(Rc<ProductionMatch<'gram>>),
+/// A [`crate::Term`] which has been "matched" while parsing input
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub(crate) enum TermMatch<'gram> {
+    /// [`crate::Term::Terminal`] which matched with a string literal
+    Terminal(&'gram str),
+    /// [`crate::Term::Nonterminal`] which was matched with a fully completed traversal
+    Nonterminal(TraversalId),
 }
 
 append_only_vec_id!(pub(crate) TraversalId);
 
-/// A step in traversing a [`crate::Grammar`]
 #[derive(Debug)]
 pub(crate) struct Traversal<'gram> {
+    pub id: TraversalId,
+    pub unmatched: &'gram [crate::Term],
     pub input_range: InputRange<'gram>,
-    pub matching: ProductionMatching<'gram>,
+    pub production_id: ProductionId,
+    from: Option<TraversalEdge<'gram>>,
 }
 
 impl<'gram> Traversal<'gram> {
-    pub fn start_production(prod: &Production<'gram>, input_range: &InputRange<'gram>) -> Self {
-        let _span = tracing::span!(tracing::Level::TRACE, "start_production").entered();
-        let input_range = input_range.after();
-        let matching = prod.start_matching();
-        Self {
-            matching,
-            input_range,
-        }
-    }
-    pub fn duplicate_key(&self) -> TraversalDuplicateKey {
-        let _span = tracing::span!(tracing::Level::TRACE, "duplicate_key").entered();
-        TraversalDuplicateKey {
-            input_range: self.input_range.offset.clone(),
-            prod_id: self.matching.prod_id,
-            matched_term_count: self.matching.matched_count(),
-        }
-    }
-    pub fn earley(&self) -> EarleyStep<'gram> {
-        let _span = tracing::span!(tracing::Level::TRACE, "earley").entered();
-        match self.matching.next() {
-            None => {
-                let prod_match = self
-                    .matching
-                    .complete()
-                    .expect("matching must be complete because no next term");
-                EarleyStep::Complete(Rc::new(prod_match))
-            }
-            Some(term) => match term {
-                Term::Nonterminal(_) => EarleyStep::Predict(term),
-                Term::Terminal(term) => EarleyStep::Scan(term),
-            },
-        }
-    }
-    pub fn match_term(&self, term_match: TermMatch<'gram>) -> Option<Self> {
-        let _span = tracing::span!(tracing::Level::TRACE, "match_term").entered();
-        let input_len = match &term_match {
-            TermMatch::Terminal(term) => term.len(),
-            TermMatch::Nonterminal(prod) => prod.input_len,
-        };
-
-        self.matching.add_term_match(term_match).map(|matching| {
-            let input_range = self.input_range.advance_by(input_len);
-
-            Self {
-                input_range,
-                matching,
-            }
-        })
+    pub fn next_unmatched(&self) -> Option<&'gram Term> {
+        self.unmatched.get(0)
     }
 }
 
-/// Key used for ignoring duplicate [`Traversal`]s
-#[derive(Debug, PartialEq, Eq, Hash)]
-pub(crate) struct TraversalDuplicateKey {
-    input_range: InputRangeOffset,
-    prod_id: ProductionId,
-    matched_term_count: usize,
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+struct TraversalEdge<'gram> {
+    pub term: TermMatch<'gram>,
+    pub parent_id: TraversalId,
 }
 
-#[derive(Debug, Default)]
-pub(crate) struct TraversalCompletionMap<'gram> {
-    incomplete: HashMap<TermCompletionKey<'gram>, Vec<TraversalId>>,
-    complete: HashMap<TermCompletionKey<'gram>, Vec<Rc<ProductionMatch<'gram>>>>,
-}
-
-impl<'gram> TraversalCompletionMap<'gram> {
-    pub fn get_incomplete(
-        &'_ self,
-        complete_traversal: &Traversal<'gram>,
-    ) -> impl Iterator<Item = TraversalId> + '_ {
-        let key = TermCompletionKey::new(
-            complete_traversal.matching.lhs,
-            complete_traversal.input_range.offset.start,
-        );
-        self.incomplete.get(&key).into_iter().flatten().cloned()
-    }
-    pub fn get_complete(
-        &'_ self,
-        matching: &'gram Term,
-        input_range: &InputRange<'gram>,
-    ) -> impl Iterator<Item = Rc<ProductionMatch<'gram>>> + '_ {
-        let key = TermCompletionKey::new(matching, input_range.offset.total_len());
-        self.complete.get(&key).into_iter().flatten().cloned()
-    }
-    pub fn insert(&mut self, traversal: &Traversal<'gram>, id: TraversalId) {
-        match traversal.matching.next() {
-            Some(Term::Terminal(_)) => {
-                // do nothing, because terminals are irrelevant to completion
-            }
-            Some(unmatched @ Term::Nonterminal(_)) => {
-                let key =
-                    TermCompletionKey::new(unmatched, traversal.input_range.offset.total_len());
-                self.incomplete.entry(key).or_default().push(id);
-            }
-            None => {
-                let key = TermCompletionKey::new(
-                    traversal.matching.lhs,
-                    traversal.input_range.offset.start,
-                );
-                let prod_match = traversal
-                    .matching
-                    .complete()
-                    .expect("matching must be complete because no next term");
-                let prod_match = Rc::new(prod_match);
-                self.complete.entry(key).or_default().push(prod_match);
-            }
-        }
-    }
-}
-
-/// Key used for "incomplete" [`Traversal`]
-#[derive(Debug, PartialEq, Eq, Hash)]
-pub(crate) struct TermCompletionKey<'gram> {
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+struct TraversalRootKey {
+    production_id: super::grammar::ProductionId,
     input_start: usize,
-    matching: &'gram Term,
 }
 
-impl<'gram> TermCompletionKey<'gram> {
-    pub fn new(matching: &'gram Term, input_start: usize) -> Self {
+type TraversalArena<'gram> = AppendOnlyVec<Traversal<'gram>, TraversalId>;
+type TreeRootMap = HashMap<TraversalRootKey, TraversalId>;
+type TreeEdgeMap<'gram> = HashMap<TraversalEdge<'gram>, TraversalId>;
+
+#[derive(Debug)]
+pub(crate) struct TraversalMatchIter<'gram, 'tree> {
+    tree: &'tree TraversalTree<'gram>,
+    current: TraversalId,
+    last: TraversalId,
+}
+
+impl<'gram, 'tree> TraversalMatchIter<'gram, 'tree> {
+    pub fn new(last: TraversalId, tree: &'tree TraversalTree<'gram>) -> Self {
+        let mut current = last;
+        while let Some(edge) = &tree.get(current).from {
+            current = edge.parent_id;
+        }
+
         Self {
-            matching,
-            input_start,
+            current,
+            tree,
+            last,
         }
     }
 }
 
-#[derive(Debug, Default)]
-pub(crate) struct TraversalQueue<'gram> {
-    arena: AppendOnlyVec<Traversal<'gram>, TraversalId>,
-    queue: VecDeque<TraversalId>,
-    completion_map: TraversalCompletionMap<'gram>,
-    processed: HashSet<TraversalDuplicateKey>,
-}
-
-impl<'gram> TraversalQueue<'gram> {
-    pub fn new(
-        grammar: &GrammarMatching<'gram>,
-        input_range: InputRange<'gram>,
-        starting_term: &'gram Term,
-    ) -> Self {
-        let _span = tracing::span!(tracing::Level::TRACE, "TraversalQueue_new").entered();
-
-        let queue = VecDeque::new();
-        let starting_traversals = grammar
-            .get_productions_by_lhs(starting_term)
-            .map(|prod| Traversal::start_production(prod, &input_range));
-
-        let mut traversal_queue = Self {
-            queue,
-            ..Default::default()
-        };
-
-        traversal_queue.extend(starting_traversals);
-
-        traversal_queue
-    }
-
-    /// Extend queue with new [`Traversal`]s. Ignores duplicates, according to [`TraversalDuplicateKey`]
-    fn extend<I>(&mut self, traversals: I)
-    where
-        I: Iterator<Item = Traversal<'gram>>,
-    {
-        let _span = tracing::span!(tracing::Level::TRACE, "TraversalQueue_extend").entered();
-        for traversal in traversals {
-            let processed_key = traversal.duplicate_key();
-
-            let is_new_traversal = self.processed.insert(processed_key);
-
-            let _new_traversal_prefix = if is_new_traversal {
-                "new traversal"
-            } else {
-                "ignored duplicate traversal"
-            };
-
-            tracing::event!(
-                tracing::Level::TRACE,
-                "{_new_traversal_prefix}: {:#?}",
-                traversal
-            );
-
-            if !is_new_traversal {
-                continue;
-            }
-
-            let id = self.arena.push(traversal);
-            self.queue.push_back(id);
-
-            let traversal = self.arena.get(id).unwrap();
-            self.completion_map.insert(traversal, id);
+impl<'gram, 'tree> Iterator for TraversalMatchIter<'gram, 'tree> {
+    type Item = &'tree TermMatch<'gram>;
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.current == self.last {
+            return None;
         }
-    }
 
-    /// Pop the next [`Traversal`] from the queue, and invoke a provided "handler" function.
-    /// Any newly created [`Traversal`] by the "handler" should be placed in the provided output buffer,
-    /// which will be added to the queue (and filtered for duplicates).
-    pub fn handle_pop<H>(&mut self, mut handler: H) -> Option<Rc<ProductionMatch<'gram>>>
-    where
-        H: FnMut(
-            TraversalId,
-            &AppendOnlyVec<Traversal<'gram>, TraversalId>,
-            &mut TraversalCompletionMap<'gram>,
-            &mut Vec<Traversal<'gram>>,
-        ) -> Option<Rc<ProductionMatch<'gram>>>,
-    {
-        let _span = tracing::span!(tracing::Level::TRACE, "TraversalQueue_handle_pop").entered();
-        let mut created = Vec::<Traversal>::new();
-
-        while let Some(id) = self.queue.pop_front() {
-            let prod_match = handler(id, &self.arena, &mut self.completion_map, &mut created);
-            self.extend(created.drain(..));
-
-            if prod_match.is_some() {
-                return prod_match;
+        let mut scan = self.last;
+        while let Some(edge) = &self.tree.get(scan).from {
+            if self.current == edge.parent_id {
+                self.current = scan;
+                return Some(&edge.term);
             }
+
+            scan = edge.parent_id;
         }
 
         None
+    }
+}
+
+#[derive(Debug, Default)]
+pub(crate) struct TraversalTree<'gram> {
+    arena: TraversalArena<'gram>,
+    tree_roots: TreeRootMap,
+    edges: TreeEdgeMap<'gram>,
+}
+
+impl<'gram> TraversalTree<'gram> {
+    pub fn get(&self, id: TraversalId) -> &Traversal<'gram> {
+        self.arena.get(id).expect("valid traversal ID")
+    }
+    pub fn get_matching(&self, id: TraversalId) -> Option<&'gram Term> {
+        self.get(id).next_unmatched()
+    }
+    pub fn get_matched(&self, id: TraversalId) -> impl Iterator<Item = &TermMatch<'gram>> {
+        TraversalMatchIter::new(id, self)
+    }
+    pub fn predict(
+        &mut self,
+        production: &Production<'gram>,
+        input_range: &InputRange<'gram>,
+    ) -> &Traversal {
+        fn inner<'gram>(
+            arena: &mut TraversalArena<'gram>,
+            tree_roots: &mut TreeRootMap,
+            input_range: &InputRange<'gram>,
+            production: &Production<'gram>,
+        ) -> TraversalId {
+            let production_id = production.id;
+            let traversal_root_key = TraversalRootKey {
+                production_id,
+                input_start: input_range.offset.total_len(),
+            };
+
+            let traversal_root = tree_roots.entry(traversal_root_key).or_insert_with(|| {
+                let traversal = arena.push_with_id(|id| Traversal {
+                    id,
+                    production_id,
+                    unmatched: &production.rhs.terms,
+                    input_range: input_range.after(),
+                    from: None,
+                });
+                traversal.id
+            });
+
+            *traversal_root
+        }
+
+        let predicted_id = inner(
+            &mut self.arena,
+            &mut self.tree_roots,
+            input_range,
+            production,
+        );
+
+        self.get(predicted_id)
+    }
+    pub fn match_term(&mut self, parent: TraversalId, term: TermMatch<'gram>) -> &Traversal {
+        fn inner<'gram>(
+            arena: &mut TraversalArena<'gram>,
+            edges: &mut TreeEdgeMap<'gram>,
+            parent: TraversalId,
+            term: TermMatch<'gram>,
+        ) -> TraversalId {
+            let parent = arena.get(parent).expect("valid parent traversal ID");
+            let input_range = match term {
+                TermMatch::Terminal(term) => parent.input_range.advance_by(term.len()),
+                TermMatch::Nonterminal(nonterminal_traversal_id) => {
+                    let nonterminal_traversal = arena
+                        .get(nonterminal_traversal_id)
+                        .expect("valid completed traversal ID");
+
+                    parent
+                        .input_range
+                        .advance_by(nonterminal_traversal.input_range.offset.len)
+                }
+            };
+
+            let parent_id = parent.id;
+            let production_id = parent.production_id;
+            let unmatched = &parent.unmatched[1..];
+            let from = TraversalEdge { term, parent_id };
+
+            let matched = edges.entry(from).or_insert_with_key(|from| {
+                let traversal = arena.push_with_id(|id| Traversal {
+                    id,
+                    production_id,
+                    unmatched,
+                    input_range: input_range.clone(),
+                    from: Some(from.clone()),
+                });
+                traversal.id
+            });
+
+            *matched
+        }
+
+        let matched_id = inner(&mut self.arena, &mut self.edges, parent, term);
+
+        self.get(matched_id)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::earley::GrammarMatching;
+    use crate::Grammar;
+
+    fn dna_grammar() -> Grammar {
+        let grammar: Grammar = "<dna> ::= <base> | <base> <dna>
+            <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
+            .parse()
+            .unwrap();
+
+        grammar
+    }
+
+    fn traversal_test_setup<'a>(
+        grammar: &'a Grammar,
+        input: &'static str,
+    ) -> (GrammarMatching<'a>, InputRange<'static>, TraversalTree<'a>) {
+        let matching = GrammarMatching::new(&grammar);
+        let input = InputRange::new(input);
+        let tree = TraversalTree::default();
+
+        (matching, input, tree)
+    }
+
+    #[test]
+    fn predict() {
+        let grammar = dna_grammar();
+        let (grammar, input, mut tree) = traversal_test_setup(&grammar, "GATTACA");
+        let production = grammar.productions_iter().next().unwrap();
+
+        let predicted = tree.predict(production, &input);
+
+        assert_eq!(&production.rhs.terms, predicted.unmatched);
+    }
+
+    #[test]
+    fn predict_again() {
+        let grammar = dna_grammar();
+        let (grammar, input, mut tree) = traversal_test_setup(&grammar, "GATTACA");
+        let production = grammar.productions_iter().next().unwrap();
+
+        let first = tree.predict(production, &input).id;
+        let again = tree.predict(production, &input).id;
+
+        assert_eq!(first, again);
+    }
+
+    #[test]
+    fn match_term() {
+        let grammar = "<start> ::= 'A'".parse().unwrap();
+        let (grammar, input, mut tree) = traversal_test_setup(&grammar, "AAAA");
+        let production = grammar.productions_iter().next().unwrap();
+        let prediction = tree.predict(production, &input).id;
+        let term_match = TermMatch::Terminal("A");
+
+        let scanned = tree.match_term(prediction, term_match);
+
+        assert_eq!(scanned.unmatched, &production.rhs.terms[1..]);
+    }
+
+    #[test]
+    fn match_term_again() {
+        let grammar = "<start> ::= 'A'".parse().unwrap();
+        let (grammar, input, mut tree) = traversal_test_setup(&grammar, "AAAA");
+        let production = grammar.productions_iter().next().unwrap();
+        let prediction = tree.predict(production, &input).id;
+
+        let term_match = TermMatch::Terminal("A");
+
+        let first = tree.match_term(prediction, term_match.clone()).id;
+        let again = tree.match_term(prediction, term_match.clone()).id;
+
+        assert_eq!(first, again);
+    }
+
+    #[test]
+    fn match_term_complete() {
+        let grammar = "<start> ::= 'A' | 'B' | 'C'".parse().unwrap();
+        let (grammar, input, mut tree) = traversal_test_setup(&grammar, "ABC");
+        let production = grammar.productions_iter().next().unwrap();
+        let prediction = tree.predict(production, &input).id;
+
+        // build match tree of <start> -> 'A', <start> -> 'B', <start> -> 'C'
+        for term_match in ["A", "B", "C"] {
+            let term_match = TermMatch::Terminal(term_match);
+            let traversal = tree.match_term(prediction, term_match);
+            assert_eq!(traversal.next_unmatched(), None);
+        }
     }
 }
